(feature) User color preferences reflected in example properly (#2)
