+++ template/src/app/navigation/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "navigation",
+  "private": true
+}
