+++ template/src/app/navigation/pages/Details/Details.tsx
@@ -4,11 +4,12 @@ import Icon from 'react-native-vector-icons/FontAwesome';
 import {useColors} from 'app/hooks/useColors';
 
 export const Details = (): JSX.Element => {
-  const {backgroundColor} = useColors();
+  const {isDarkMode, backgroundColor} = useColors();
+  const isAndroid = Platform.OS === 'android';
 
   return (
-    <View style={[styles.background, Platform.OS === 'android' && {backgroundColor}]}>
-      <Icon name='rocket' size={30} color='#900' />
+    <View style={[styles.background, isAndroid && {backgroundColor}]}>
+      <Icon name='rocket' size={30} color={isDarkMode && isAndroid ? 'white' : '#900'} />
       <Text>If you see a rocket, everything is working!</Text>
     </View>
   );
