+++ template/src/app/navigation/pages/Details/package.json
