+++ template/src/app/navigation/pages/Home/package.json
