+++ template/src/app/navigation/pages/package.json
